Some J5 migration and spotless (#3494)

+++ src/net/sourceforge/kolmafia/persistence/CandyDatabase.java
@@ -62,7 +62,7 @@ public class CandyDatabase {
     List<AdventureResult> others = new ArrayList<>();
 
     for (Integer itemId : CandyDatabase.tier2Candy) {
-      AdventureResult item = ItemPool.get(itemId.intValue());
+      AdventureResult item = ItemPool.get(itemId);
       if (ConsumablesDatabase.getFullness(item.getName()) > 0) {
         foods.add(item);
       } else if (ItemDatabase.isPotion(item)) {
